requirements.txt
crewai==1.9.3
crewai_tools==1.9.3
langchain_community==0.4.1
langchain_groq==1.1.2
langchain_huggingface==1.2.0
langchain_text_splitters==1.1.0
python-dotenv==1.2.1
